pyproject.toml
[project]
name = "docs2md"
version = "1.0.0"
description = "A robust Python tool that converts various document formats (PDF, Word, Excel, Images, Audio, HTML, PowerPoint, text-based files like CSV/JSON/XML, ZIP, YouTube URLs, EPUBs) to Markdown format with customizable post-processing adjustments."
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "mjanez", email = "96422458+mjanez@users.noreply.github.com"},
]

dependencies = [
    "python-docx>=1.2.0",
    "markitdown[all]>=0.1.2",
    "openai>=1.58.1",
    "pyyaml==6.0"
]

requires-python = "==3.11.*"

[tool.pdm]
distribution = false

[project.urls]
"Bug Tracker" = "https://github.com/mjanez/docs2md/issues"
repository = "https://github.com/mjanez/docs2md"